Rename installTemplate into createFromTemplate
