+++ lib/plugins/create/create.js
@@ -7,7 +7,7 @@ const _ = require('lodash');
 const untildify = require('untildify');
 
 const ServerlessError = require('../../classes/Error').ServerlessError;
-const installTemplate = require('../../utils/installTemplate');
+const createFromTemplate = require('../../utils/createFromTemplate');
 const userStats = require('../../utils/userStats');
 const download = require('../../utils/downloadTemplateFromRepo');
 const renameService = require('../../utils/renameService').renameService;
@@ -226,7 +226,7 @@ class Create {
       this.serverless.config.update({ servicePath: process.cwd() });
     }
 
-    return installTemplate(this.options.template, process.cwd()).then(
+    return createFromTemplate(this.options.template, process.cwd()).then(
       () => {
         // rename the service if the user has provided a path via options and is creating a service
         if ((boilerplatePath || serviceName) && notPlugin) {
