+++ lib/plugins/interactiveCli/initializeService.js
@@ -3,7 +3,7 @@
 const { join } = require('path');
 const chalk = require('chalk');
 const inquirer = require('inquirer');
-const installTemplate = require('../../utils/installTemplate');
+const createFromTemplate = require('../../utils/createFromTemplate');
 const {
   getConfigFilePath,
   getServerlessConfigFile,
@@ -68,7 +68,7 @@ module.exports = {
           }
           return projectNameInput(workingDir).then(projectName => {
             const projectDir = join(workingDir, projectName);
-            return installTemplate(projectType, projectDir)
+            return createFromTemplate(projectType, projectDir)
               .then(() => {
                 process.stdout.write(
                   `\n${chalk.green(`Project successfully created in '${projectName}' folder.`)}\n`
