+++ lib/utils/createFromTemplate.js
