+++ lib/utils/createFromTemplate.test.js
@@ -3,14 +3,14 @@
 const path = require('path');
 const chai = require('chai');
 const { existsSync } = require('fs-extra');
-const installTemplate = require('./installTemplate');
+const installTemplate = require('./createFromTemplate');
 const { getTmpDirPath } = require('../../tests/utils/fs');
 
 chai.use(require('chai-as-promised'));
 
 const expect = chai.expect;
 
-describe('#installTemplate()', () => {
+describe('#createFromTemplate()', () => {
   let tmpDirPath;
 
   beforeEach(() => {
@@ -18,7 +18,7 @@ describe('#installTemplate()', () => {
     return installTemplate('aws-nodejs', tmpDirPath);
   });
 
-  it('should install template', () =>
+  it('should create from template', () =>
     expect(existsSync(path.join(tmpDirPath, 'serverless.yml'))).to.be.true);
 
   it('should handle .gitignore', () =>
